CBOR: split up encoder and decoder
